test(webapp): split esbuild live-wasm tests from mocked loader tests
The file-wide vi.mock intercepted getEsbuild for the SLICC_TEST_HEAVY_WASM
opt-in suite too, defeating its integration check. Move the live tests into
an unmocked file and share the ctx factory via a helper.
